style(runbotics-orchestrator-ui): :lipstick: Redesign of language switcher
