+++ runbotics/runbotics-orchestrator-ui/src/src-app/hooks/useClickOutsideComponent.ts
@@ -1,0 +1,17 @@
+import {useEffect, MutableRefObject} from 'react';
+
+function useClickOutsideComponent(ref: MutableRefObject<any> , handler: Function) {
+    useEffect(() => {
+        function handleClickOutside(event) {
+            if (ref.current && !ref.current.contains(event.target)) {
+                handler();
+            }
+        }
+        document.addEventListener('mousedown', handleClickOutside);
+        return () => {
+            document.removeEventListener('mousedown', handleClickOutside);
+        };
+    }, [ref]);
+}
+
+export default useClickOutsideComponent;
